+++ src/uncrustify.cpp
@@ -995,7 +995,7 @@ static int load_mem_file(const char *filename, file_mem &fm)
                  fm.enc == char_encoding_e::e_UTF8 ? "UTF-8" :
                  fm.enc == char_encoding_e::e_UTF16_LE ? "UTF-16-LE" :
                  fm.enc == char_encoding_e::e_UTF16_BE ? "UTF-16-BE" : "Error",
-                 fm.enc);
+                 (int)fm.enc);
          retval = 0;
       }
    }
